Make vue-config builder
